Move registry to models/
