+++ lib/models/registry.js
@@ -1,9 +1,9 @@
 'use strict';
 
-var Plugin           = require('./plugin');
-var StylePlugin      = require('./style-plugin');
-var TemplatePlugin   = require('./template-plugin');
-var JavascriptPlugin = require('./javascript-plugin');
+var Plugin           = require('../preprocessors/plugin');
+var StylePlugin      = require('../preprocessors/style-plugin');
+var TemplatePlugin   = require('../preprocessors/template-plugin');
+var JavascriptPlugin = require('../preprocessors/javascript-plugin');
 var debug            = require('debug')('ember-cli:registry');
 
 function Registry(plugins, app) {
