+++ lib/preprocessors.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var path         = require('path');
-var Registry     = require('./preprocessors/registry');
+var Registry     = require('./models/registry');
 var requireLocal = require('./utilities/require-local');
 
 function setupRegistryForEachAddon(registry, parent) {
